Add files via upload
The beginning of a first draft of a book on Surface Algebras. May be useful for understanding the varieties of complexes, some equivariant neural network layers, K-theoretic ideas. The derived categories and cluster categories are work that is not currently included but that others such as Paquette, Plamondon, Keller, Liu, Amiot, Bauer, and Schroll have been working on for quite some time. I was not able to ever dig deeply into this beyond a few basic results on almost split sequences implied by results of [Representation Theory of an Infinite Quiver](https://arxiv.org/abs/1109.3176) and [On the Auslander-Reiten quiver of the representations of an infinite quiver](https://arxiv.org/abs/1201.4833).